+++ packages/common/utils/src/check-value-is-filled.ts
@@ -1,0 +1,7 @@
+import { Any } from '@angular-ru/common/typings';
+
+import { checkValueIsEmpty } from './check-value-is-empty';
+
+export function checkValueIsFilled(value: Any): boolean {
+    return !checkValueIsEmpty(value);
+}